Fix for issue #335
Created seperate files for each example
