+++ examples/ips/README.md
@@ -1,0 +1,5 @@
+![SendGrid Logo](https://uiux.s3.amazonaws.com/2016-logos/email-logo%402x.png)
+
+This folder contains various examples on using the IPs endpoint of SendGrid with Java:
+
+* [Retrieve all IP addresses (GET /ips)](retrieveallips.java)
